+++ Runtime/Events/EventTuple.cs
@@ -1,4 +1,5 @@
 using System;
+using TriInspector;
 using UnityEngine;
 
 namespace GameUtils
@@ -6,8 +7,8 @@ namespace GameUtils
     [Serializable]
     public class EventTuple
     {
-        public GameObject Caller;
-        public string ClassName;
-        public string MethodName;
+        [SerializeField, Group("Caller"), HideLabel] public GameObject CallerGameObject;
+        [SerializeField, Group("Caller Script"), HideLabel] public MonoBehaviour CallerScript;
+        [SerializeField, Group("Method Name"), HideLabel] public string MethodName;
     }
 }