+++ Runtime/Events/GameEventAsset.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using TriInspector;
+using UnityEngine;
 
 namespace GameUtils
 {
@@ -24,29 +25,30 @@ namespace GameUtils
             _callHistory = new List<T>();
         }
 
-        public void AddListener(Action<T> action)
+        public void AddListener(MonoBehaviour caller, Action<T> action)
         {
             if (action == null)
+            {
+                this.LogWarning($"Attempted to add a null listener.", this);
                 return;
+            }
 
             //
-            /*_runtimeListeners.Add(new EventTuple
+            _runtimeListeners.Add(new EventTuple
             {
-                Caller = ,
-                ClassName = action.Method.DeclaringType?.Name,
+                CallerGameObject = caller.gameObject,
+                CallerScript = caller,
                 MethodName = action.Method.Name
-            });*/
+            });
 
             //
             _onInvoked += action;
         }
 
-        public void RemoveListener(Action<T> action)
+        public void RemoveListener(MonoBehaviour caller, Action<T> action)
         {
             // Deletes the listener and its reference from the runtime listeners list.
-            //_runtimeListeners.RemoveAll(tuple => tuple.Caller == action.Target?.ToString() && tuple.MethodName == action.Method.Name);
-
-            //
+            _runtimeListeners.RemoveAll(tuple => tuple.CallerScript == caller && tuple.MethodName == action.Method.Name);
             _onInvoked -= action;
         }
 
@@ -57,6 +59,7 @@ namespace GameUtils
             _onInvoked = null;
         }
 
+        [Button(ButtonSizes.Medium)]
         public virtual void Invoke(T param)
         {
             // Gestisce l'invocazione dell'evento con tracciamento opzionale.